+++ src/test/java/com/javabom/baseballgame/domain/domain/number/ClientGameNumbersTest.java
@@ -1,6 +1,7 @@
 package com.javabom.baseballgame.domain.domain.number;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
@@ -8,8 +9,33 @@ import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ClientGameNumbersTest {
+
+    @DisplayName("게임에 사용될 숫자의 갯수가 3이 아니면 IllegalArgumentException을 발생시킨다.")
+    @Test
+    void checkCount() {
+        //given
+        List<String> strClientGameNumber = Arrays.asList("1", "2");
+
+        //then
+        assertThatThrownBy(() -> new ClientGameNumbers(strClientGameNumber))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(String.format("%d, 숫자 게임에서 사용될 숫자의 갯수는 3개여야 합니다.", strClientGameNumber.size()));
+    }
+
+    @DisplayName("게임에 사용될 숫자에 중복이 있을 경우 IllegalArgumentException을 발생시킨다.")
+    @Test
+    void checkDuplicate() {
+        //given
+        List<String> strClientGameNumber = Arrays.asList("1", "2", "2");
+
+        //then
+        assertThatThrownBy(() -> new ClientGameNumbers(strClientGameNumber))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(String.format("중복을 제외한 숫자 갯수: %d, 숫자에 중복이 있습니다.", 2));
+    }
 
     @DisplayName("들어온 숫자의 순서와 숫자 둘 같으면 true, 하나라도 다르면 false를 반환한다.")
     @ParameterizedTest
